Converted experience to long, updated the migrator
Former-commit-id: 3b5b713a4bf363a386f8d06d3b2cfb91e7b34720

+++ Intersect Library/Intersect Library/GameObjects/ClassBase.cs
@@ -9,7 +9,7 @@ namespace Intersect.GameObjects
         public int AttackAnimation = -1;
 
         //Exp Calculations
-        public int BaseExp = 100;
+        public long BaseExp = 100;
 
         public int BasePoints;
         public int[] BaseStat = new int[(int) Stats.StatCount];
@@ -23,7 +23,7 @@ namespace Intersect.GameObjects
         public int Damage;
 
         public int DamageType;
-        public int ExpIncrease = 50;
+        public long ExpIncrease = 50;
 
         //Level Up Info
         public int IncreasePercentage;
@@ -132,8 +132,8 @@ namespace Intersect.GameObjects
             PointIncrease = myBuffer.ReadInteger();
 
             //Exp Info
-            BaseExp = myBuffer.ReadInteger();
-            ExpIncrease = myBuffer.ReadInteger();
+            BaseExp = myBuffer.ReadLong();
+            ExpIncrease = myBuffer.ReadLong();
 
             //Regen
             for (int i = 0; i < (int) Vitals.VitalCount; i++)
@@ -217,8 +217,8 @@ namespace Intersect.GameObjects
             myBuffer.WriteInteger(PointIncrease);
 
             //Exp Info
-            myBuffer.WriteInteger(BaseExp);
-            myBuffer.WriteInteger(ExpIncrease);
+            myBuffer.WriteLong(BaseExp);
+            myBuffer.WriteLong(ExpIncrease);
 
             //Regen
             for (int i = 0; i < (int) Vitals.VitalCount; i++)
