+++ Intersect Library/Intersect Library/GameObjects/NpcBase.cs
@@ -21,7 +21,7 @@ namespace Intersect.GameObjects
         //Drops
         public List<NPCDrop> Drops = new List<NPCDrop>();
 
-        public int Experience;
+        public long Experience;
         public int Level = 1;
 
         //Vitals & Stats
@@ -71,7 +71,7 @@ namespace Intersect.GameObjects
             {
                 Stat[i] = myBuffer.ReadInteger();
             }
-            Experience = myBuffer.ReadInteger();
+            Experience = myBuffer.ReadLong();
             SpawnDuration = myBuffer.ReadInteger();
             Behavior = myBuffer.ReadByte();
             SightRange = myBuffer.ReadInteger();
@@ -124,7 +124,7 @@ namespace Intersect.GameObjects
             {
                 myBuffer.WriteInteger(Stat[i]);
             }
-            myBuffer.WriteInteger(Experience);
+            myBuffer.WriteLong(Experience);
             myBuffer.WriteInteger(SpawnDuration);
             myBuffer.WriteByte(Behavior);
             myBuffer.WriteInteger(SightRange);
